Separate analyzers
