+++ src/unimport/analyzers/__init__.py
@@ -1,0 +1,3 @@
+from src.unimport.analyzers.main import MainAnalyzer
+
+__all__ = ("MainAnalyzer",)
